# Faster float formatting