… ext/json/ext/vendor/fpconv/README.md
… ext/json/ext/vendor/fpconv/license
… ext/json/ext/vendor/fpconv/src/fpconv.c
… ext/json/ext/vendor/fpconv/src/fpconv.h
… ext/json/ext/vendor/fpconv/src/powers.h